🛡️ fix(config): fail closed on registry and credentials
